refactor: test structure + process
- before each test, copy included ui5 app into tmp path
- prep server config files into tmp path
- run ui5 app from tmp path
- assert http responses
